Renamed CodeCamperDataService to CodeCamperUow.
Refactored several custom repository members from property form (a mistake) to Get... methods because they are queries and should be methods (not properties) named in the same manner as other repository queries (i.e., with a "Get" prefix).

Added the PersonsRepository and relocated Sessions.GetSpeakers to Persons.GetSpeakers; adjusted controllers accordingly.

Broke out the repository factory aspects of the RepositoryProvider into a "RepositoryFactories" class. The factories should only be defined once for the server application (in a "singleton" class); RepositoryProviders, which deliver repository instances on demand, are created once for each request. The different lifetimes of the two concerns motivated this refactoring.

There is no longer a need for a CodeCamperRepositoryProvider but I had to add a CodeCamperRepositoryFactories class which defines the custom repository factories for Code Camper.