Revised model.mapper.js to serve "mapper" objects which have more than one mapping function; datacontext.js modified to suit.
I did this because I "needed" two mapping functions: a function to map from dto's to model objects AND a function to get the model id from the dto.

That didn’t used to be an issue when we could assume that every dto had an "Id" property. But Attendance does not.

At first I faked an Attendance.Id property on the server; I added a readonly Id property and serialized it. I also had to tell EF to ignore this property as it is not mapped to the db. This approach seemed wrong and I backed it out in this commit; the server-side Attendance is restored to its former self. That restoration meant I had to construct the Id for the dto on the client side. Thus the need for two mapper functions per type … and thus the “mapper” object.

Adapted the datacontext.js to the mapper objects.

An incidental virtue of the mapper object is that it can support any number of mapping capabilities including "toDto" which might be needed to strip out the extra UI-oriented properties that we add to the client models. Such cruft is presently shipped to the server on save.

The cruft isn't a problem today … the Web API model binding seems to ignore it. But these non-persisted properties could become a problem someday; we should not assume (as we presently do) that every property of the model object could be serialized and sent to the server. If we got into trouble, we’d need a "toDto" function to serialize just the essential properties.