Updated NuGet.exe to address parallel package restore problem.
We lean on NuGet package restore to get the packages from the web. So, upon first build, VS tries to get them. But a VS2012 perf improvement is to process projects async if it can. This can cause multiple NuGet restores which conflicted … until you update NuGet.exe ... as I did here.