Removed trailing commas that caused IE7/8 to fail. (model.js and dataservice.attendance.js)
Added require('datacontext') to the model.js to help with circular reference.

Changed main.js to use require('boostrapper').run()

Added missing dependency for amplify.request on jquery in main.js

Forced parameters for the plugins in the bottom of main.js (in requirejs function). I think this solves the problem where plugins sometimes do not load. If not, let's remove it later.