Before, I had to use ko.validaiton to extend the models. which i liked doing.  I then added required and regex matches to the models.
But if i want to use html5 validation attributes and css3 validation selectors to make those rbowsers shine too, i had to manually add them to the html. so it was done in 2 places. which sucks since it can get out of synch.

I asked Eric Barnard to add them to the html if it was supported. he did, as an option for bwards compat. i tested it and patterns did not work ... but he fixed that in a second push. we both tested and it looks good now.

So in html5 browsers, it gets the color coding styles and the ko.validation message and in non html5 browsers it just gets ko.validation message